fix(shell): canonicalize audience route
