+++ apps/web/tests/e2e/smoke-auth.spec.ts
@@ -78,7 +78,7 @@ async function assertDashboardRouteLoaded(
     return;
   }
 
-  if (path === APP_ROUTES.DASHBOARD_AUDIENCE) {
+  if (path === APP_ROUTES.AUDIENCE) {
     await expect(
       page.getByTestId('dashboard-audience-client'),
       'Audience: audience client did not render'
@@ -230,7 +230,7 @@ test.describe('Dashboard Navigation @smoke', () => {
       ? [{ path: APP_ROUTES.CHAT, name: 'Chat' }]
       : [
           { path: APP_ROUTES.CHAT, name: 'Chat' },
-          { path: APP_ROUTES.DASHBOARD_AUDIENCE, name: 'Audience' },
+          { path: APP_ROUTES.AUDIENCE, name: 'Audience' },
           { path: APP_ROUTES.DASHBOARD_RELEASES, name: 'Releases' },
           {
             path: APP_ROUTES.DASHBOARD_EARNINGS,
