+++ apps/web/tests/e2e/smoke-prod-auth.spec.ts
@@ -264,7 +264,7 @@ test.describe('Production Auth Smoke @production-smoke', () => {
 
     await waitForHydration(page);
 
-    const tabs = [APP_ROUTES.DASHBOARD_AUDIENCE, APP_ROUTES.DASHBOARD_RELEASES];
+    const tabs = [APP_ROUTES.AUDIENCE, APP_ROUTES.DASHBOARD_RELEASES];
 
     for (const tabPath of tabs) {
       await page.goto(tabPath, {
