+++ apps/web/tests/global-setup.ts
@@ -152,7 +152,7 @@ async function globalSetup() {
     APP_ROUTES.CHAT, // auth.setup.ts navigates here — warm up to avoid cold-compile 404
     APP_ROUTES.LIBRARY,
     APP_ROUTES.DASHBOARD_RELEASES,
-    APP_ROUTES.DASHBOARD_AUDIENCE,
+    APP_ROUTES.AUDIENCE,
     APP_ROUTES.PRESENCE,
     APP_ROUTES.EARNINGS,
     APP_ROUTES.DASHBOARD_PROFILE,
