+++ apps/web/tests/unit/dashboard/DashboardAudienceTable.test.tsx
@@ -39,7 +39,7 @@ vi.mock('next/navigation', () => ({
     prefetch: vi.fn(),
   }),
   useSearchParams: () => new URLSearchParams(),
-  usePathname: () => '/app/dashboard/audience',
+  usePathname: () => '/app/audience',
 }));
 
 vi.mock('@/contexts/TableMetaContext', () => ({
